refactor(prompt): rename build_enhanced_rag_prompt to build_rag_prompt and remove dead code - Rename build_enhanced_rag_prompt -> build_rag_prompt in enhanced_prompt_builder.py - Delete prompt_builder.py (old baseline, now dead code) - Update send_message.py imports and calls - Inline old baseline prompt in benchmark tests to preserve before/after comparison
