+++ server/src/raggae/application/use_cases/chat/send_message.py
@@ -43,7 +43,7 @@ from raggae.domain.exceptions.project_exceptions import (
     ProjectReindexInProgressError,
 )
 from raggae.domain.value_objects.organization_member_role import OrganizationMemberRole
-from raggae.infrastructure.services.enhanced_prompt_builder import build_enhanced_rag_prompt
+from raggae.infrastructure.services.enhanced_prompt_builder import build_rag_prompt
 
 _API_KEY_PROVIDERS = {"openai", "gemini", "anthropic"}
 
@@ -239,7 +239,7 @@ class SendMessage:
             history_window_size=effective_history_window_size,
             history_max_chars=effective_history_max_chars,
         )
-        prompt = build_enhanced_rag_prompt(
+        prompt = build_rag_prompt(
             query=message,
             context_chunks=[chunk.content for chunk in relevant_chunks],
             source_filenames=[chunk.document_file_name or "" for chunk in relevant_chunks],
@@ -449,7 +449,7 @@ class SendMessage:
             history_window_size=effective_history_window_size,
             history_max_chars=effective_history_max_chars,
         )
-        prompt = build_enhanced_rag_prompt(
+        prompt = build_rag_prompt(
             query=message,
             context_chunks=[chunk.content for chunk in relevant_chunks],
             source_filenames=[chunk.document_file_name or "" for chunk in relevant_chunks],
